move VocalUtilities into tools package
